✨[Feature] ENUM 수정 완료 및 테스트 코드 수정 완료
